requirements.txt
absl-py==0.11.0
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
comtypes==1.1.7
docopt==0.6.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
Js2Py==0.70
Keras-Preprocessing==1.1.2
Markdown==3.3.3
mysql-connector==2.2.9
mysql-connector-python==8.0.22
num2words==0.5.10
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pipwin==0.5.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Mala%20Jha/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pyjsparser==2.7.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
pyttsx3==2.90
pytz==2020.4
pywin32==300
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
word2number==1.1
wrapt==1.12.1
zipp==3.4.0
